Finalised File Structure with working headers being produced
