--- RPI/voltage.csv
@@ -1,7 +1,0 @@
-Name:;Default;ID:;f6940638-f7eb-48cb-94f0-24c5d0d22f64;Time Interval;0
-0A0;0A1
-Edit Me;Edit Me
-1;1
-0;0
-0;0
-Edit Me 2;Edit Me 2
